installer: chromeos-setgoodkernel: Abort if FAFT lockfile exists
This script can mess up FAFT testing by changing things in the GPT
or in the firmware while FAFT is explicitly testing bad behaviors.

FAFT tests will add a lockfile which is cleaned up upon success
and this file can check for the existing lockfile and abort early
if it exists.

BUG=chrome-os-partner:30832
TEST=tested with pending changes to FAFT to create a lockfile on
test startup and remove it on test cleanup to ensure that this
script does not run during FAFT tests.

Change-Id: I580447772ed05ebdde65b67d6be9ec6e0497756c
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222001
Reviewed-by: Bill Richardson <wfrichar@chromium.org>