init: if boot splash fails, trigger udev events
If the boot splash fails, the most likely reason is the graphics driver
has not been built into the kernel. The udev daemon is responsible for
loading them, so this patch triggers udev to load them very early so
that the boot splash will function. This is only triggered in the case
where the boot splash fails, so this adds negligible boot time on
systems where the boot splash works the first time.

BUG=chromium:420907
TEST=run a test image under qemu and check that boot splash runs

Change-Id: Ie52a0f18578883721906681a2098a600a1086d7b
Reviewed-on: https://chromium-review.googlesource.com/222332
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Commit-Queue: Zach Reizner <zachr@chromium.org>