shill: Device: Destroy dangling connection
When we disconnect, the refcounted connection_ object in the
device must be destroyed.  This object's destructor removes
routing table entries and IP addresses.  Before this CL, the
connectivity tester created dangling references to this
connection, which caused a left-over connection object to
remain even after the system had re-connected.

This dangling connection would only finally be removed when
a new connectivity report was created in
Device::StartConnectivityTest(), which would then remove the
(current) connection state from the device, rendering the
system routeless and IP-less.

This CL removes the old connectivity tester at the same time
as other objects that reference the connection (like the
portal detector), and adds a unit test to make sure that
the connection is dereferenced when DestroyConnection is
called.

BUG=chromium:421722
TEST=New unit test + manual:
 - Connect to a wireless network
 - Browse to "chome://system" (triggering a connectivity test)
 - Disconnect and re-connect to the network
 - Refresh the "chrome://system" table (re-triggering the test)
 - Ensure the system still has connectivity

Change-Id: I94c316f89bd3f0dc9c1c11cfa0e8bb197653d415
Reviewed-on: https://chromium-review.googlesource.com/222458
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>