power: suspend_stress_test: Use wakeup_count
This patch updates the suspend_stress_test script (which still seems to
be in widespread use despite my valiant efforts to kill it :( ) to pass
the --wakeup_count flag to powerd_dbus_suspend, like the autotests
already did for a long time. This is necessary to prevent powerd from
retrying a suspend that failed due to a concurrent wake event, leading
to a stuck-in-suspend machine.

However, with that fix we would now silently ignore any concurrent wake
events. Since they can still often be a symptom for real issues, instead
change our failure detection to count successes instead of failures.
This way every suspend that never even started is counted as a failure
(and we can also get rid of the explicit "suspend command failure"
detection).

BUG=chrome-os-partner:32848
TEST=Ran on Nyan_Blaze (both normal and --memory_suspend), confirmed it
still works and counts failures correctly. Mashed key during suspend to
trigger concurrent wakeup, confirmed that it was correctly detected as a
failure and that powerd didn't retry the suspend.

Change-Id: I605d97d2ca8aa21403c9e30f2613b625fb55b850
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222881
Reviewed-by: Daniel Erat <derat@chromium.org>