login: Emit PropertyChanged signal only after changing config.
This way, it is clear that consumers of the PropertyChangeComplete
signal can assume that the side effects have taken place. Note that the
code re-ordering is only for clarity, because EmitSignal only queues a
task that will be executed in a subsequent message loop iteration.
Better to not implicitly rely on that though.

BUG=chromium:422948
TEST=platform_SessionManagerBlockDevmodeSetting

Change-Id: Id2fd5ceeb085a6ad330ae6ba084adeb33d78a14e
Reviewed-on: https://chromium-review.googlesource.com/223102
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Mattias Nissler <mnissler@chromium.org>