login: Add methods for registering start/finish of supervised user creation
Chrome can't tolerate the browser crashing and restarting during the
creation of a new supervised user account. If this happens, the user
winds up signed into the custodian account, which is undesirable.

Provide methods for the browser to inform the session_manager that
this is going on, so it can take appropriate action.

CQ-DEPEND=CL:222916
BUG=chromium:422996
TEST=unit

Change-Id: Ief82c03e1744ff2390d56b4351e5d8dd1d650f3e
Reviewed-on: https://chromium-review.googlesource.com/222972
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>