login: Add to_be_frozen cgroup sub-directory
The chrome_renderers freezer cgroup is used to freeze renderers for
lucid sleep.  However, we don't want to freeze those renderers that host
GCM apps and extensions since they are the ones that will be doing the
actual work.  This adds a new sub-group to the chrome_renderers freezer
cgroup so that we can track both the renderers that should be frozen and
the ones that should not.

BUG=chromium:364339
TEST=manual
Deploy on samus.  stop ui.  start ui.  Check that the directory exists.

Change-Id: I807264607e124e36e443438dbaa02c814b579880
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/223367