libwebserv: Add simple reference web server interface library
NOTE: This is just a reference implementation. The design
of web server implementation and the daemon-facing interface
library is still in progress. However a working webserver
in Chrome OS core is needed to unblock other teams. This
temporary implementation will be revisited once the overall
design is approved.

Added a simple web server interface implementation based on
libmicrohttpd. The server can be launched and it can start
responding to HTTP connections sent to the specified port.

Request handlers can be registered for particular URL paths
and process requests...

BUG=chromium:422029
TEST=Compiles and runs. Manual testing on device
CQ-DEPEND=CL:222888

Change-Id: I078bd75917292c72fa45cce4045a27ae6fc6b564
Reviewed-on: https://chromium-review.googlesource.com/220862
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Vitaly Buka <vitalybuka@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>