login: Return error when RestartJob() call is disallowed.
The session_manager performs an extra check on the PID of the
caller when someone tries to call RestartJob() over DBus. It
used to simply drop the incoming call when it decided the call
was disallowed, but this leads to callers (in tests, for example)
having to wait for the bus to time the call out. Returning an
error is just as correct and more friendly for integration tests.

BUG=chromium:427111
TEST=security_RestartJob should finish about 25s more quickly.

Change-Id: I73699f82ed982b41b4a204ac54d26c1ad0b37c8a
Reviewed-on: https://chromium-review.googlesource.com/226373
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>