easy-unlock: Add a method that converts public key to format used by EasyUnlock
As part of easy signin protocol, Chrome has to pass a TPM public key to the
Android phone (incorporated into a challenge sent to the phone). The phone
expects the key to be passed as a securemessage GenericPublicKey proto.
TPM public key used by Chrome is in DER format, so it has to be converted
to the right format. Because Chrome itself does not yet have securemessage
support, the conversion logic is added to the easy-unlock daemon.

BUG=chromium:409027
TEST=FEATURES="test" emerge-link easy-unlock
cbuildbot link-paladin x86-generic-paladin daisy-paladin

CQ-DEPEND=CL:227154

Change-Id: Iddb8c8075449e1e7c3e08046110c1aa607d5274b
Reviewed-on: https://chromium-review.googlesource.com/226956
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Tested-by: Toni Barzic <tbarzic@chromium.org>