peerd: Allow setting the initial mDNS prefix via the commandline
This allows us to test our behavior when we encounter prefix collisions.
We need to set this to a known value, because mDNS lets existing
services continue to use a conflicting name, and specifies that the
newcomer must change their own name.  This means to cause a conflict we
had to have set up a service with a known prefix before starting peerd.

BUG=chromium:426173
TEST=Integration tests validate that this works as intended.

Change-Id: Ie3f3d25adf003e33c561b48765e075acb90f18c0
Reviewed-on: https://chromium-review.googlesource.com/226642
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>