peerd: Correct Manager.StartMonitoring signature
Change the list of technologies to be a list of rather than magic
numbers.  This is more legible for debugging.

Add a dictionary of string to variant values to enable future passing
of options to StartMonitoring.  Currently we support no options.

Add a property org.chromium.peerd.Manager.MonitoredTechnologies that
exposes of a list of the supported technologies we're
currently monitoring for peers to interested parties.

BUG=chromium:430185
TEST=unittests, updated integration tests pass.

Change-Id: I593ec8480d11c290c5a3e37670f849b24c289715
Reviewed-on: https://chromium-review.googlesource.com/227463
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>