login: Exit gracefully if Owner key is corrupted.
In the event that the session_manager discovers that the
Owner key is corrupt, it triggers a device wipe. It used
to crash, which led to the UI getting restarted a bunch
and then the device rebooting. If we instead exit gracefully,
then the device should reboot to wipe immediately.

BUG=chromium:429656
TEST=unit

Change-Id: Id5238bd52c1eb8bf827216e7c7be81620880f274
Reviewed-on: https://chromium-review.googlesource.com/227511
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>