power: WA apply brightness value after toggling display power
The bl_power sysfs is asynchronous and can cause it to race with
dpms toggling.  When transitioning from off, this may cause the
dpms operation to interleave with the bl_power operation in some
drivers and result in the backlight not getting set back to it's
original value.

The dpms operation should block powerd until it finishes.  Thus
applying the brightness value afterwards should avoid the potential
for interleaving.

BUG=chrome-os-partner:29857,chrome-os-partner:31186
TEST=Connect HDMI, then while the eDP panel is transitioning from
     dim->off, quickly move the mouse pointer to cause transition
     from off->normal event.  eDP panel brightness should return
     to original value (pre-dim value).

Change-Id: Ic8ec0097547d97c8d5880001be10a37b6257db07
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/228311
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Matthew E Hiltner <matthew.e.hiltner@intel.com>