login: Ensure that DBus call to reboot goes through
When the session_manager discovers corruption of the Owner public key,
it initiates a powerwash. It needs to tell powerd to reboot the device
in order to do do. Apparently, the DBus method call was being queued
but not sent before, though the device would indeed eventually reboot
due to the UI continually exiting quickly. Now, the session_manager
blocks on the call being sent, ensuring that the reboot will happen.

BUG=chromium:429656
TEST=unit tests pass, also tested manually

Change-Id: I327357b9cd2392a3a8bfb404f79e28fe73bd67ce
Reviewed-on: https://chromium-review.googlesource.com/228875
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>