libwebserv: Fix a crash in web server when doing some POST requests
The main issue was that Response was holding on to a reference to
shared pointer to the connection, not the shared pointer itself.
This was fixed and the signature of method handler was switched to
scoped_ptr to emphasize the ownership transfer.

However, switching from shared_ptr to scoped_ptr caused another
not so desirable side effect since there is no specialization 
for scoped_ptr containing  objects (like the request parameter 
in the handler). So I had make it non-const (which sucks).

Also changed libwebserv::Connection to be base::RefCounted to play
nicely with base::Bind/base::Callback too.

There were two additional issues discovered:
1. Request::GetHeader() had infinite loop (Oops!)
2. Apparently sending a response through libmicrohttpd should be
   in response of its callback. So, Response::SendResponse now
   hands the data to the Connection class which feeds it back
   to libmicrohttp when it calls back through
   ServerHelper::ConnectionHandler().

BUG=chromium:431164
TEST=FEATURES=test emerge-x86-alex libwebserv privetd
     Manually tested on a device.

Change-Id: I0f1cf32da56db1aeb05935efcff0109bba78dad5
Reviewed-on: https://chromium-review.googlesource.com/228950
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Vitaly Buka <vitalybuka@chromium.org>
Commit-Queue: Vitaly Buka <vitalybuka@chromium.org>
Tested-by: Vitaly Buka <vitalybuka@chromium.org>