power: Set up dark resume on request, not attempt
This shuffles around the time at which we schedule
dark resumes, and makes the dark resume callback responsible
for scheduling itself in the future. Callbacks are descheduled
at the end of a request (i.e. on user resume). This is based
on the patch sketch Chirantan provided in CL 228948.

This hopefully simplifies the logic in the DarkResume class,
even if the test code gets a bit more complicated since we have
to keep track of what counts as a "user resume" (and when to
"fire" the RTC alarm).

This supersedes CLs 228361, 228866, 228867, and 228868.

BUG=chrome-os-partner:31970
TEST=manual, unit

Change-Id: I6449d895846037335cba3b319d3fee60c0d67467
Reviewed-on: https://chromium-review.googlesource.com/230343
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>