libchromeos: Add TouchFile() function.
Add a function to touch a file, creating it if it doesn't exist. Several
similar functions exist throughout platform2, the goal here is to
provide a single function that can be used instead.

This implementation of touch aims to be a little more powerful by removing any
directories that may be blocking the file, creating any needed parent
directories, and setting new file permissions.

BUG=chromium:403170
TEST=Ran included unittests, manual testing.

Change-Id: Iddaf2cd9546b89ffdab593c952934e048bfabcbf
Reviewed-on: https://chromium-review.googlesource.com/231259
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David Pursell <dpursell@chromium.org>
Tested-by: David Pursell <dpursell@chromium.org>