libchromeos: Fix GPU acceleration on freon boards
Currently we are only able to obtain the gbm path on two boards, link
and peppy. In particular, because the dri ozone platform is defined,
the --disable-gpu flag is added to the Chrome command line. If we want
to scale convertiong boards to freon, we need something which scales
and therefore doesn't require adding board names to a long if()
statement. This change fixes this by removing the test for link/peppy.

BUG=chromium:428125
TEST=build auron with freon changes, boot it, get GPU acceleration

Change-Id: Ia2590c9ca3489fd7d84052144419c667d71ff442
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/231213