Revert "Revert "privetd: Add HTTPS support""
Putting back the original commit from CL:230920 that was reverted
in 5a89228b1f7f9a253ede66cefe1b590651afc697.

The problem on ASan builders was that the private key and the
certificate buffers were binary blobs that were passed to
libmicrohttpd which expected them to be zero-terminated strings.
It was doing strlen on those buffers and going past their
boundaries, which was happily killed by ASan's buffer checks.

Here is the description of the original CL for reference:

This is the initial implementation of HTTPS support on privetd
using the temporary web server solution based on libmicrohttpd.

Currently, privetd daemon generates X509 certificate each time
it starts and does not save it anywhere. This will need to be
changed to make sure the paired clients can continue to trust
the device after it has been rebooted.

BUG=chromium:434916
TEST=FEATURES=test emerge-link libwebserv privetd
     Manually test on device through a broweser
     Updated privetd_WebServerSanity autotest which passes on
     both HTTP and HTTPS
CQ-DEPEND=CL:231234

Change-Id: I75d27177c181ff565aa1bf9846afaaa834607783
Reviewed-on: https://chromium-review.googlesource.com/231233
Reviewed-by: Vitaly Buka <vitalybuka@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Vitaly Buka <vitalybuka@chromium.org>